+++ src/Entity/VariableFeesLine.php
@@ -66,12 +66,12 @@ class VariableFeesLine
         return $this;
     }
 
-    public function getAmount(): ?int
+    public function getAmount(): ?float
     {
         return $this->amount;
     }
 
-    public function setAmount(int $amount): self
+    public function setAmount(float $amount): self
     {
         $this->amount = $amount;
